start clear command

+++ Deployment/0c00d01.ps1
@@ -1,36 +1,37 @@
 # Deploy CDB Commands...slot environment and region is covered by CDB redundancy mechanisms
-param ($tenant='apico', $set='ba', $project='rt', $service='co', $version='00', $lane='1', $defaultregion='s')
+param ($tenant = 'apico', $set = 'ba', $project = 'rt', $service = 'co', $version = '00', $lane = '1', $defaultregion = 's')
 
-$slot='0'
-$environment='0'
-$region='0'
+$slot = '0'
+$environment = '0'
+$region = '0'
 
 $objecttype = '0c'
 $operation = '00'
 $area = 'd'
 $name = `
-    -join($tenant,$set,$project,$service,$version,$objecttype,$operation,$area,'c','0',$lane,$slot,$environment,$region)
+    -join ($tenant, $set, $project, $service, $version, $objecttype, $operation, $area, 'c', '0', $lane, $slot, $environment, $region)
 $database = `
-    -join($tenant,$set,$project,$service,$version,$objecttype,$operation,$area,'d','0',$lane,$slot,$environment,$region)
+    -join ($tenant, $set, $project, $service, $version, $objecttype, $operation, $area, 'd', '0', $lane, $slot, $environment, $region)
 $collection = `
-    -join($tenant,$set,$project,$service,$version,$objecttype,$operation,$area,'o','0',$lane,$slot,$environment,$region)
+    -join ($tenant, $set, $project, $service, $version, $objecttype, $operation, $area, 'o', '0', $lane, $slot, $environment, $region)
 $group = `
-    -join($tenant,$set,$project,$service,$version,$objecttype,$operation,$area,'g','0',$lane,$slot,$environment,$region)
+    -join ($tenant, $set, $project, $service, $version, $objecttype, $operation, $area, 'g', '0', $lane, $slot, $environment, $region)
 
-$location = switch($defaultregion) {
-    's' {'Southeast Asia'; break} 
-    'e' {'East Asia'; break} 
- }
+$location = switch ($defaultregion) {
+    's' { 'Southeast Asia'; break } 
+    'e' { 'East Asia'; break } 
+}
 
- $template=-join('00','00','0','c','0','.template.json')
+$template = -join ('00', '00', '0', 'c', '0', '.template.json')
 
+# not in production !
+if ($lane -ne 'z') {
+    Write-Host "======================="
+    Write-Host "Delete RG Commands Data"
+    Write-Host "======================="
 
-Write-Host "======================="
-Write-Host "Delete RG Commands Data"
-Write-Host "======================="
-
-Remove-AzResourceGroup -Name $group -Force -ErrorAction SilentlyContinue
-
+    Remove-AzResourceGroup -Name $group -Force -ErrorAction SilentlyContinue
+}
 
 Write-Host "======================="
 Write-Host "Create RG Commands Data"
@@ -42,4 +43,4 @@ Write-Host "========================"
 Write-Host "Create CDB Commands Data"
 Write-Host "========================"
 
-New-AzResourceGroupDeployment -ResourceGroupName $group -TemplateFile $template -TemplateParameterObject @{ name = $name;  location = $location; database = $database; collection = $collection; }
+New-AzResourceGroupDeployment -ResourceGroupName $group -TemplateFile $template -TemplateParameterObject @{ name = $name; location = $location; database = $database; collection = $collection; }
